drm: Boost DDR bus upon running an atomic ioctl
In order to reduce jank, request a DDR bus boost whenever an atomic
ioctl is sent, which is what our display HAL uses to commit frames to
the display. The boost should be sufficient enough to render the
display's maximum FPS without any dropped frames when there is no
significant external source of load.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: If9d114a8058e0ff4835dd2c3d04a915377ee6c53