wakeup_stats: Don't export device info to sysfs
wakeup_source_device_create creates a symlink to the directory of the
device associated with the wakelock. Not only is this unnecessary, this
also results in selinux denials and the following errors in logcat if
the directory isn't whitelisted in sepolicy:

"Error opening kernel wakelock stats for:
 wakeup[n]: Permission denied"

Stop creating the symlink to silence the errors as it's useless anyway.

Signed-off-by: Kazuki Hashimoto <kaz205@tuta.io>
Signed-off-by: baalajimaestro <me@baalajimaestro.me>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
(cherry picked from commit eef1a65b14df489bc91cec8994d4c08d1a3ef58f)
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>